Refactored button code
